+++ packages/frontend/src/hellfall/hooks/useSearchResults.ts
@@ -50,8 +50,7 @@ import {
   colorMiscReduce,
   hybridColorCompOp,
   hybridIdentityMiscReduce,
-  numCompOp,
-} from '../filters/opComps.ts';
+} from '../filters/filterColors.ts';
 import { textEquals, textSearchIncludes } from '@hellfall/shared/utils/textHandling.ts';
 import { CHUNK_SIZE } from '../constants.ts';
 import { extraSetList } from '@hellfall/shared/data/sets.ts';
@@ -59,6 +58,7 @@ import { filterSet } from '../filters/filterSet.ts';
 import { filterText, filterTextList } from '../filters/filterText.ts';
 import { looseOpType, opType } from '../filters/types.ts';
 import { getAllNames } from '../getNames.ts';
+import { filterNumber, filterNumberString } from '../filters/filterNumber.ts';
 
 export const useSearchResults = () => {
   const location = useLocation();
@@ -277,44 +277,44 @@ export const useSearchResults = () => {
           return false;
         }
         if (collectorNumber) {
-          if (!numCompOp(entry.collector_number, collectorNumber[1], collectorNumber[0])) {
+          if (!filterNumberString(entry.collector_number, collectorNumber[1], collectorNumber[0])) {
             return false;
           }
         }
         if (manaValue) {
-          if (!numCompOp(entry.mana_value, manaValue[1], manaValue[0])) {
+          if (!filterNumberString(entry.mana_value, manaValue[1], manaValue[0])) {
             return false;
           }
         }
         if (power) {
-          if (!numCompOp(entry.toFaces()[0].power, power[1], power[0])) {
+          if (!filterNumberString(entry.toFaces()[0].power, power[1], power[0])) {
             return false;
           }
         }
         if (toughness) {
-          if (!numCompOp(entry.toFaces()[0].toughness, toughness[1], toughness[0])) {
+          if (!filterNumberString(entry.toFaces()[0].toughness, toughness[1], toughness[0])) {
             return false;
           }
         }
         if (loyalty) {
-          if (!numCompOp(entry.toFaces()[0].loyalty, loyalty[1], loyalty[0])) {
+          if (!filterNumberString(entry.toFaces()[0].loyalty, loyalty[1], loyalty[0])) {
             return false;
           }
         }
         if (defense) {
-          if (!numCompOp(entry.toFaces()[0].defense, defense[1], defense[0])) {
+          if (!filterNumberString(entry.toFaces()[0].defense, defense[1], defense[0])) {
             return false;
           }
         }
 
         if (colorNumber) {
-          if (!numCompOp(entry.colors.length, colorNumber[1], colorNumber[0])) {
+          if (!filterNumber(entry.colors.length, colorNumber[1], colorNumber[0])) {
             return false;
           }
         }
         if (colorIdentityNumber) {
           if (
-            !numCompOp(entry.color_identity.length, colorIdentityNumber[1], colorIdentityNumber[0])
+            !filterNumber(entry.color_identity.length, colorIdentityNumber[1], colorIdentityNumber[0])
           ) {
             return false;
           }
