+++ packages/shared/src/data/tags.json
@@ -16,7 +16,6 @@
 		"2015-card-frame",
 		"2015-frame",
 		"2015-token-frame",
-		"2020-card-frame",
 		"2020-token-frame",
 		"2b2t",
 		"2dreadmaw",
