+++ packages/shared/src/data/tokens.json
@@ -103590,6 +103590,182 @@
 		},
 		{
 			"object": "card",
+			"id": "sans eheheh1",
+			"name": "sans eheheh",
+			"set": "HCT",
+			"collector_number": "1585",
+			"layout": "token",
+			"image_status": "highres",
+			"image": "https://lh3.googleusercontent.com/d/18gMhX_SPLTdZheO85iavTOgVoDwritY0",
+			"mana_cost": "",
+			"mana_value": 0,
+			"types": [
+				"Creature"
+			],
+			"subtypes": [
+				"Skeleton"
+			],
+			"type_line": "Creature — Skeleton",
+			"oracle_text": "Flying\\nWhen sans eheheh enters, you may exile target nonland permanent an opponent controls. For as long as that card remains exiled, its owner may cast it if they descended this turn.",
+			"power": "1",
+			"toughness": "1",
+			"colors": [
+				"B"
+			],
+			"color_identity": [],
+			"color_identity_hybrid": [],
+			"keywords": [],
+			"legalities": {
+				"standard": "not_legal",
+				"4cb": "not_legal",
+				"commander": "not_legal"
+			},
+			"creators": [
+				"sjonesy1"
+			],
+			"rulings": "",
+			"border_color": "white",
+			"frame": "jank",
+			"frame_effects": [
+				"inverted"
+			],
+			"tags": [
+				"white-border",
+				"inverted-text",
+				"jank-frame"
+			],
+			"variation": false,
+			"isActualToken": true,
+			"all_parts": [
+				{
+					"object": "related_card",
+					"id": "5041",
+					"name": "Angel of Undertale",
+					"set": "HC7.1",
+					"image": "https://lh3.googleusercontent.com/d/1-iOJCucoBajD65s2_8tO4MclT0wakiUE",
+					"type_line": "Creature — Angel",
+					"component": "token_maker"
+				}
+			]
+		},
+		{
+			"object": "card",
+			"id": "Shoe1",
+			"name": "Shoe",
+			"set": "HCT",
+			"collector_number": "1586",
+			"layout": "token",
+			"image_status": "highres",
+			"image": "https://lh3.googleusercontent.com/d/1JLQN3IjmvXriCPClQPM_UypzC2VCrMKp",
+			"mana_cost": "",
+			"mana_value": 0,
+			"types": [
+				"Artifact"
+			],
+			"subtypes": [
+				"Equipment"
+			],
+			"type_line": "Artifact — Equipment",
+			"oracle_text": "Equipped creature gets +0/+1.\\nEquip {1}",
+			"flavor_text": "\"These shoes rule.\"\\n\"These shoes suck.\"",
+			"colors": [],
+			"color_identity": [],
+			"color_identity_hybrid": [],
+			"keywords": [],
+			"legalities": {
+				"standard": "not_legal",
+				"4cb": "not_legal",
+				"commander": "not_legal"
+			},
+			"creators": [
+				"multiloquent"
+			],
+			"rulings": "",
+			"border_color": "black",
+			"frame": "token_2015",
+			"tags": [
+				"2015-frame"
+			],
+			"variation": false,
+			"isActualToken": true,
+			"all_parts": [
+				{
+					"object": "related_card",
+					"id": "1250",
+					"name": "Falling Head // Heels",
+					"set": "HC2.1",
+					"image": "https://lh3.googleusercontent.com/d/1NgNOc20g510yJonQryIDRo4V_K56GVSK",
+					"type_line": "Sorcery // Sorcery",
+					"component": "token_maker"
+				}
+			]
+		},
+		{
+			"object": "card",
+			"id": "Mummifire1",
+			"name": "Mummifire",
+			"set": "HCT",
+			"collector_number": "1587",
+			"layout": "token",
+			"image_status": "highres",
+			"image": "https://lh3.googleusercontent.com/d/1RTMXOGs_IwuawrSyOmbwyTxEJQLgpSFe",
+			"mana_cost": "",
+			"mana_value": 0,
+			"types": [
+				"Creature"
+			],
+			"subtypes": [
+				"Zombie"
+			],
+			"type_line": "Creature — Zombie",
+			"oracle_text": "Mummifire deals 2 damage to any target. (This happens only when this instant resolves.)\\nWhenever Mummifire enters or attacks, cast a copy of it. (The instant version.)",
+			"power": "4",
+			"toughness": "4",
+			"colors": [
+				"B"
+			],
+			"color_identity": [],
+			"color_identity_hybrid": [],
+			"keywords": [],
+			"legalities": {
+				"standard": "not_legal",
+				"4cb": "not_legal",
+				"commander": "not_legal"
+			},
+			"creators": [
+				"multiloquent"
+			],
+			"rulings": "",
+			"border_color": "black",
+			"frame": "2015",
+			"frame_effects": [
+				"showcase",
+				"inverted"
+			],
+			"tags": [
+				"2015-card-frame",
+				"showcase-frame",
+				"inverted-text"
+			],
+			"tag_notes": {
+				"showcase-frame": "Signature Spellbook"
+			},
+			"variation": false,
+			"isActualToken": true,
+			"all_parts": [
+				{
+					"object": "related_card",
+					"id": "1278",
+					"name": "Mummifire",
+					"set": "HC2.1",
+					"image": "https://lh3.googleusercontent.com/d/1ncz_L9eVZ2E2Y0zBT0o7q4WH2MpINgk8",
+					"type_line": "Instant",
+					"component": "token_maker"
+				}
+			]
+		},
+		{
+			"object": "card",
 			"id": "Day // Night1",
 			"scryfall_id": "dc26e13b-7a0f-4e7f-8593-4f22234f4517",
 			"oracle_id": "84142407-ba98-497d-bbc8-39176aa67c02",
